Created a SpawnManager object, this object is the parent of the spawners (any new spawners created need to have the SpawnManager as their parent for spawning to work properly).
The Spawn Manager has a script which simply keeps track of how many enemies there are.
Whenever a Spawner creates an enemy, it adds 1 to the SpawnManagers numofEnemies variable.
Whenever an Enemy is destroyed, it removes 1 from the numOfEnemies.
The Spawner will not spawn units unless the global numOfEnemies is less than its spawnCap variable.