ITS THURSDAY, so you know what that means, time to do a whole weeks work in 6 hours.
Juicy Juicy firing has been added, YAY!!!!

- Added the BulletTrigger Class, this will detect collsions with object and destroy the bullet (unless it collides with the player).
If it collides with an enemy it applies some damage to the enemy.
- Tweaked the rather PlayerProto class to support firing at the mouse position,
Added a bunch of necessary variable, including a fireRate, which affects how many shots are fired per second.
Also added firing to the Controller's Right Stick Axis', If both the mouse is on screen and the controller stick axis is not at 0, the game will prioritise the Right Stick Axis for aiming.

-Can't remember what else I added, derp.